+++ extensions/bitwarden/src/api.ts
@@ -1,23 +1,20 @@
-import { getPreferenceValues } from "@raycast/api";
 import execa from "execa";
 import { existsSync } from "fs";
 import { dirname } from "path/posix";
-import { PasswordGeneratorOptions, VaultStatus } from "./types";
+import { Item, PasswordGeneratorOptions, VaultState } from "./types";
 import { getPasswordGeneratingArgs } from "./utils";
 
 export class Bitwarden {
   private env: Record<string, string>;
-  private cliPath: string;
-  constructor() {
-    const { cliPath, clientId, clientSecret } = getPreferenceValues();
-    if (cliPath) {
-      this.cliPath = cliPath;
-    } else {
-      this.cliPath = process.arch == "arm64" ? "/opt/homebrew/bin/bw" : "/usr/local/bin/bw";
+  cliPath: string;
+  constructor(clientId: string, clientSecret: string, cliPath: string) {
+    if (!cliPath) {
+      cliPath = process.arch == "arm64" ? "/opt/homebrew/bin/bw" : "/usr/local/bin/bw";
     }
-    if (!existsSync(this.cliPath)) {
-      throw Error(`Invalid Cli Path: ${this.cliPath}`);
+    if (!existsSync(cliPath)) {
+      throw new Error(`Bitwarden CLI not found at ${cliPath}`);
     }
+    this.cliPath = cliPath;
     this.env = {
       BW_CLIENTSECRET: clientSecret.trim(),
       BW_CLIENTID: clientId.trim(),
@@ -33,9 +30,15 @@ export class Bitwarden {
     await this.exec(["login", "--apikey"]);
   }
 
-  async listItems<ItemType>(type: string, sessionToken: string): Promise<ItemType[]> {
-    const { stdout } = await this.exec(["list", type, "--session", sessionToken]);
-    return JSON.parse(stdout);
+  async logout(): Promise<void> {
+    await this.exec(["logout"]);
+  }
+
+  async listItems(sessionToken: string): Promise<Item[]> {
+    const { stdout } = await this.exec(["list", "items", "--session", sessionToken]);
+    const items = JSON.parse(stdout);
+    // Filter out items without a name property (they are not displayed in the bitwarden app)
+    return items.filter((item: Item) => !!item.name);
   }
 
   async getTotp(id: string, sessionToken: string): Promise<string> {
@@ -53,9 +56,9 @@ export class Bitwarden {
     await this.exec(["lock"]);
   }
 
-  async status(sessionToken: string | undefined): Promise<VaultStatus> {
-    const { stdout } = await this.exec(sessionToken ? ["status", "--session", sessionToken] : ["status"]);
-    return JSON.parse(stdout).status;
+  async status(): Promise<VaultState> {
+    const { stdout } = await this.exec(["status"]);
+    return JSON.parse(stdout);
   }
 
   async generatePassword(options?: PasswordGeneratorOptions): Promise<string> {
@@ -65,6 +68,6 @@ export class Bitwarden {
   }
 
   private async exec(args: string[]): Promise<execa.ExecaChildProcess> {
-    return execa(this.cliPath, args, { env: this.env });
+    return execa(this.cliPath, args, { env: this.env, input: "" });
   }
 }