+++ extensions/gif-search/src/models/gif.ts
@@ -7,6 +7,21 @@ export interface IGif {
   slug: string;
   preview_gif_url: string;
   gif_url: string;
+  metadata?: {
+    width?: number;
+    height?: number;
+    size?: number;
+    labels?: {
+      title: string;
+      text: string;
+    }[];
+    links?: {
+      title: string;
+      target: string;
+      text: string;
+    }[];
+    tags?: string[];
+  };
   attribution?:
     | {
         dark: string;
@@ -17,19 +32,14 @@ export interface IGif {
 
 export function renderGifMarkdownDetails(gif: IGif) {
   let md = `
-## ${gif.title}
+### ${gif.title}
 
-![${gif.title}](${gif.gif_url})
+<img alt="${gif.title}" src="${gif.gif_url}" height="200" />
 
-\`\`\`
-Static preview, animated preview coming soon!
-\`\`\`
   `;
 
   if (gif.attribution) {
-    md += `
-![Powered by](file:${environment.assetsPath}/${gif.attribution})
-`;
+    md += `<img height="36" alt="Powered by" src="file:${environment.assetsPath}/${gif.attribution}" />`;
   }
 
   return md;