+++ extensions/safari/src/utils.ts
@@ -3,7 +3,7 @@ import _ from "lodash";
 import osascript from "osascript-tag";
 import { URL } from "url";
 
-import { showToast, Toast } from "@raycast/api";
+import { Icon, showToast, Toast } from "@raycast/api";
 
 import { HistoryItem, Tab } from "./types";
 
@@ -58,9 +58,14 @@ export const getUrlDomain = (url: string) => {
 };
 
 export const getFaviconUrl = (domain: string | undefined) => {
-  if (domain) {
-    return `https://www.google.com/s2/favicons?sz=64&domain=${encodeURI(domain)}`;
+  if (!domain) {
+    return Icon.Globe;
   }
+
+  return {
+    source: `https://www.google.com/s2/favicons?sz=64&domain=${encodeURI(domain)}`,
+    fallback: Icon.Globe,
+  };
 };
 
 export const formatDate = (date: string) =>