+++ extensions/todo-list/src/open_url_action.tsx
@@ -2,11 +2,22 @@ import { Action, Keyboard } from "@raycast/api";
 import { URL } from "url";
 
 const OpenUrlAction = ({ url, shortcut, title }: { url: string; shortcut?: Keyboard.Shortcut; title?: string }) => {
-  const hostname = new URL(url).hostname;
+  let hostname;
+  let resolvedUrl = url;
+  try {
+    hostname = new URL(url).hostname;
+  } catch (e) {
+    try {
+      resolvedUrl = `https://${url}`;
+      hostname = new URL(resolvedUrl).hostname;
+    } catch (e) {
+      return null;
+    }
+  }
   const faviconUrl = `https://www.google.com/s2/favicons?sz=64&domain=${hostname}`;
   return (
     <>
-      <Action.OpenInBrowser icon={{ source: faviconUrl }} shortcut={shortcut} title={title ?? url} url={url} />
+      <Action.OpenInBrowser icon={{ source: faviconUrl }} shortcut={shortcut} title={title ?? url} url={resolvedUrl} />
     </>
   );
 };