+++ extensions/sentry/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Sentry Changelog
 
+## Accessories Support (2022-04-06)
+
+- Support new accessories in the List.Item for sentry issues
+
 ## Initial Version (2022-03-09)
 
 - Search unresolved issues by title and filter by project